requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pandas==2.2.0
openai==1.12.0
python-dotenv==1.0.1
pydantic-settings==2.1.0
openpyxl==3.1.2
requests